working on http integration

+++ Java project/src/main/java/ee/ut/cs/dsg/confcheck/trie/TrieNode.java
@@ -18,7 +18,7 @@ public class TrieNode {
     private List<Integer> linkedTraces;
     private int level=0;
     private int numChildren=0;
-    private int confidenceCost;
+    private float confidenceCost;
 
     private double scaledConfCost;
 
@@ -247,11 +247,12 @@ public class TrieNode {
         else
             return  child;
     }
-    public int getConfidenceCost() {
+
+    public float getConfidenceCost() {
         return confidenceCost;
     }
 
-    public void setConfidenceCost(int confidenceCost) {
+    public void setConfidenceCost(float confidenceCost) {
         this.confidenceCost = confidenceCost;
     }
